Downloads/virus_hybrid_deployment/requirements.txt
streamlit
torch
torchvision
opencv-python-headless
pillow
numpy
tifffile